% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workout3.R
\name{bin_cumulative}
\alias{bin_cumulative}
\title{bin_cumulative}
\usage{
bin_cumulative(trials, prob)
}
\arguments{
\item{trials}{number of fixed trials}

\item{prob}{probability of success on each trial}
}
\value{
data frame with success, probability and cumulative sections
}
\description{
calculates the distribution of probability over a number of successes and the cumulative probability
}
\examples{
bin_cumulative(trials = 5, prob = 0.5)
bin_cumulative(5, 0.3)
}
