% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workout3.R
\name{bin_kurtosis}
\alias{bin_kurtosis}
\title{bin_kurtosis}
\usage{
bin_kurtosis(trials, prob)
}
\arguments{
\item{trials}{number of fixed trials}

\item{prob}{probability of success on each trial}
}
\value{
the kurtosis
}
\description{
main function that computes kurtosis using trials and probability
}
\examples{
bin_kurtosis(trials = 10, prob = 0.3)
bin_kurtosis(10, 0.3)
}
