% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workout3.R
\name{bin_mean}
\alias{bin_mean}
\title{bin_mean}
\usage{
bin_mean(trials, prob)
}
\arguments{
\item{trials}{number of fixed trials}

\item{prob}{probability of success on each trial}
}
\value{
the mean
}
\description{
main function that computes mean using trials and probability
}
\examples{
bin_mean(trials = 10, prob = 0.5)
bin_mean(10, 0.3)
}
