% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workout3.R
\name{bin_mode}
\alias{bin_mode}
\title{bin_mode}
\usage{
bin_mode(trials, prob)
}
\arguments{
\item{trials}{number of fixed trials}

\item{prob}{probability of success on each trial}
}
\value{
the mode
}
\description{
main function that computes the mode using trials and probability
}
\examples{
bin_mode(trials = 10, prob = 0.3)
bin_mode(10, 0.3)
}
