% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workout3.R
\name{bin_probability}
\alias{bin_probability}
\title{bin_probability}
\usage{
bin_probability(success, trials, prob)
}
\arguments{
\item{success}{the number of successes from trials}

\item{trials}{number of fixed trials}

\item{prob}{probability of success on each trial}
}
\value{
integer that represents probability
}
\description{
calculates the probability distribution using trials, success and probability
}
\examples{
bin_probability(success = 2, trials = 5, prob = 0.5)
bin_probability(0:2, 5, 0.5)
}
