% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workout3.R
\name{bin_skewness}
\alias{bin_skewness}
\title{bin_skewness}
\usage{
bin_skewness(trials, prob)
}
\arguments{
\item{trials}{number of fixed trials}

\item{prob}{probability of success on each trial}
}
\value{
the skewness
}
\description{
main function that finds skewness of distribution using trials and probability
}
\examples{
bin_skewness(trials = 10, prob = 0.3)
bin_skewness(10, 0.3)
}
