% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workout3.R
\name{bin_variable}
\alias{bin_variable}
\title{bin_variable}
\usage{
bin_variable(trials, prob)
}
\arguments{
\item{trials}{number of fixed trials}

\item{prob}{probability of success on each trial}
}
\value{
a binomial random variable object, "binvar"
}
\description{
lists trials and probability in a binomial random variable object
}
\examples{
bin_variable(trials = 5, prob = 0.5)
bin_variable(5, 0.3)
}
