% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workout3.R
\name{bin_variance}
\alias{bin_variance}
\title{bin_variance}
\usage{
bin_variance(trials, prob)
}
\arguments{
\item{trials}{number of fixed trials}

\item{prob}{probability of success on each trial}
}
\value{
the variance
}
\description{
main function that computes variance using trials and probability
}
\examples{
bin_variance(trials = 10, prob = 0.5)
bin_variance(10, 0.5)
}
